# colexec: add default comparison operator