--- a/pkg/sql/colexec/colbuilder/execplan.go
+++ b/pkg/sql/colexec/colbuilder/execplan.go
@@ -1456,38 +1456,43 @@ func planSelectionOperators(
 		}
 		lTyp := ct[leftIdx]
 		if constArg, ok := t.Right.(tree.Datum); ok {
-			if t.Operator == tree.Like || t.Operator == tree.NotLike {
-				negate := t.Operator == tree.NotLike
+			switch cmpOp {
+			case tree.Like, tree.NotLike:
+				negate := cmpOp == tree.NotLike
 				op, err = colexec.GetLikeOperator(
-					evalCtx, leftOp, leftIdx, string(tree.MustBeDString(constArg)), negate)
-				return op, resultIdx, ct, internalMemUsedLeft, err
-			}
-			if t.Operator == tree.In || t.Operator == tree.NotIn {
-				negate := t.Operator == tree.NotIn
+					evalCtx, leftOp, leftIdx, string(tree.MustBeDString(constArg)), negate,
+				)
+			case tree.In, tree.NotIn:
+				negate := cmpOp == tree.NotIn
 				datumTuple, ok := tree.AsDTuple(constArg)
 				if !ok {
-					err = errors.Errorf("IN is only supported for constant expressions")
-					return nil, resultIdx, ct, internalMemUsed, err
+					// Optimized IN operator is supported only on constant
+					// expressions, so we fallback to the default comparison
+					// operator.
+					break
 				}
 				op, err = colexec.GetInOperator(lTyp, leftOp, leftIdx, datumTuple, negate)
-				return op, resultIdx, ct, internalMemUsedLeft, err
-			}
-			if t.Operator == tree.IsDistinctFrom || t.Operator == tree.IsNotDistinctFrom {
-				if t.Right != tree.DNull {
-					err = errors.Errorf("IS DISTINCT FROM and IS NOT DISTINCT FROM are supported only with NULL argument")
-					return nil, resultIdx, ct, internalMemUsed, err
+			case tree.IsDistinctFrom, tree.IsNotDistinctFrom:
+				if constArg != tree.DNull {
+					// Optimized IsDistinctFrom and IsNotDistinctFrom are
+					// supported only with NULL argument, so we fallback to the
+					// default comparison operator.
+					break
 				}
 				// IS NOT DISTINCT FROM NULL is synonymous with IS NULL and IS
 				// DISTINCT FROM NULL is synonymous with IS NOT NULL (except for
 				// tuples). Therefore, negate when the operator is IS DISTINCT
 				// FROM NULL.
-				negate := t.Operator == tree.IsDistinctFrom
+				negate := cmpOp == tree.IsDistinctFrom
 				op = colexec.NewIsNullSelOp(leftOp, leftIdx, negate)
-				return op, resultIdx, ct, internalMemUsedLeft, err
 			}
-			op, err := colexec.GetSelectionConstOperator(
-				lTyp, t.TypedRight().ResolvedType(), cmpOp, leftOp, leftIdx, constArg, nil, /* binFn */
-			)
+			if op == nil {
+				// op hasn't been created yet, so let's try the constructor for
+				// all other selection operators.
+				op, err = colexec.GetSelectionConstOperator(
+					cmpOp, leftOp, ct, leftIdx, constArg, evalCtx, t,
+				)
+			}
 			return op, resultIdx, ct, internalMemUsedLeft, err
 		}
 		rightOp, rightIdx, ct, internalMemUsedRight, err := planProjectionOperators(
@@ -1497,7 +1502,7 @@ func planSelectionOperators(
 			return nil, resultIdx, ct, internalMemUsed, err
 		}
 		op, err := colexec.GetSelectionOperator(
-			lTyp, ct[rightIdx], cmpOp, rightOp, leftIdx, rightIdx, nil, /* binFn */
+			cmpOp, rightOp, ct, leftIdx, rightIdx, evalCtx, t,
 		)
 		return op, resultIdx, ct, internalMemUsedLeft + internalMemUsedRight, err
 	default:
@@ -1563,7 +1568,7 @@ func planProjectionOperators(
 	case *tree.ComparisonExpr:
 		return planProjectionExpr(
 			ctx, evalCtx, t.Operator, t.ResolvedType(), t.TypedLeft(), t.TypedRight(),
-			columnTypes, input, acc, factory, nil, /* binFn */
+			columnTypes, input, acc, factory, nil /* binFn */, t,
 		)
 	case *tree.BinaryExpr:
 		if err = checkSupportedBinaryExpr(t.TypedLeft(), t.TypedRight(), t.ResolvedType()); err != nil {
@@ -1571,7 +1576,7 @@ func planProjectionOperators(
 		}
 		return planProjectionExpr(
 			ctx, evalCtx, t.Operator, t.ResolvedType(), t.TypedLeft(), t.TypedRight(),
-			columnTypes, input, acc, factory, t.Fn,
+			columnTypes, input, acc, factory, t.Fn.Fn, nil, /* cmpExpr */
 		)
 	case *tree.IsNullExpr:
 		t.TypedInnerExpr()
@@ -1782,7 +1787,8 @@ func planProjectionExpr(
 	input colexecbase.Operator,
 	acc *mon.BoundAccount,
 	factory coldata.ColumnFactory,
-	binFn *tree.BinOp,
+	binFn tree.TwoArgFn,
+	cmpExpr *tree.ComparisonExpr,
 ) (op colexecbase.Operator, resultIdx int, typs []*types.T, internalMemUsed int, err error) {
 	if err := checkSupportedProjectionExpr(left, right); err != nil {
 		return nil, resultIdx, typs, internalMemUsed, err
@@ -1806,8 +1812,8 @@ func planProjectionExpr(
 		// The projection result will be outputted to a new column which is appended
 		// to the input batch.
 		op, err = colexec.GetProjectionLConstOperator(
-			colmem.NewAllocator(ctx, acc, factory), left.ResolvedType(), typs[rightIdx], outputType,
-			projOp, input, rightIdx, lConstArg, resultIdx, binFn, evalCtx,
+			colmem.NewAllocator(ctx, acc, factory), typs, left.ResolvedType(), outputType,
+			projOp, input, rightIdx, lConstArg, resultIdx, evalCtx, binFn, cmpExpr,
 		)
 	} else {
 		var (
@@ -1826,36 +1832,44 @@ func planProjectionExpr(
 			// The projection result will be outputted to a new column which is appended
 			// to the input batch.
 			resultIdx = len(typs)
-			if projOp == tree.Like || projOp == tree.NotLike {
+			switch projOp {
+			case tree.Like, tree.NotLike:
 				negate := projOp == tree.NotLike
 				op, err = colexec.GetLikeProjectionOperator(
 					colmem.NewAllocator(ctx, acc, factory), evalCtx, input, leftIdx, resultIdx,
 					string(tree.MustBeDString(rConstArg)), negate,
 				)
-			} else if projOp == tree.In || projOp == tree.NotIn {
+			case tree.In, tree.NotIn:
 				negate := projOp == tree.NotIn
 				datumTuple, ok := tree.AsDTuple(rConstArg)
 				if !ok {
-					err = errors.Errorf("IN operator supported only on constant expressions")
-					return nil, resultIdx, typs, internalMemUsed, err
+					// Optimized IN operator is supported only on constant
+					// expressions, so we fallback to the default comparison
+					// operator.
+					break
 				}
 				op, err = colexec.GetInProjectionOperator(
 					colmem.NewAllocator(ctx, acc, factory), typs[leftIdx], input, leftIdx,
 					resultIdx, datumTuple, negate,
 				)
-			} else if projOp == tree.IsDistinctFrom || projOp == tree.IsNotDistinctFrom {
+			case tree.IsDistinctFrom, tree.IsNotDistinctFrom:
 				if right != tree.DNull {
-					err = errors.Errorf("IS DISTINCT FROM and IS NOT DISTINCT FROM are supported only with NULL argument")
-					return nil, resultIdx, typs, internalMemUsed, err
+					// Optimized IsDistinctFrom and IsNotDistinctFrom are
+					// supported only with NULL argument, so we fallback to the
+					// default comparison operator.
+					break
 				}
 				// IS NULL is replaced with IS NOT DISTINCT FROM NULL, so we want to
 				// negate when IS DISTINCT FROM is used.
 				negate := projOp == tree.IsDistinctFrom
 				op = colexec.NewIsNullProjOp(colmem.NewAllocator(ctx, acc, factory), input, leftIdx, resultIdx, negate)
-			} else {
+			}
+			if op == nil {
+				// op hasn't been created yet, so let's try the constructor for
+				// all other projection operators.
 				op, err = colexec.GetProjectionRConstOperator(
-					colmem.NewAllocator(ctx, acc, factory), typs[leftIdx], right.ResolvedType(), outputType,
-					projOp, input, leftIdx, rConstArg, resultIdx, binFn, evalCtx,
+					colmem.NewAllocator(ctx, acc, factory), typs, right.ResolvedType(), outputType,
+					projOp, input, leftIdx, rConstArg, resultIdx, evalCtx, binFn, cmpExpr,
 				)
 			}
 		} else {
@@ -1873,8 +1887,8 @@ func planProjectionExpr(
 			internalMemUsed += internalMemUsedRight
 			resultIdx = len(typs)
 			op, err = colexec.GetProjectionOperator(
-				colmem.NewAllocator(ctx, acc, factory), typs[leftIdx], typs[rightIdx], outputType,
-				projOp, input, leftIdx, rightIdx, resultIdx, binFn, evalCtx,
+				colmem.NewAllocator(ctx, acc, factory), typs, outputType, projOp,
+				input, leftIdx, rightIdx, resultIdx, evalCtx, binFn, cmpExpr,
 			)
 		}
 	}