--- a/pkg/sql/colexec/colbuilder/execplan.go
+++ b/pkg/sql/colexec/colbuilder/execplan.go
@@ -1438,13 +1438,17 @@ func planSelectionOperators(
 		op, resultIdx, typs, internalMemUsed, err = planProjectionOperators(
 			ctx, evalCtx, t.TypedInnerExpr(), columnTypes, input, acc, factory,
 		)
-		op = colexec.NewIsNullSelOp(op, resultIdx, false)
+		op = colexec.NewIsNullSelOp(
+			op, resultIdx, false /* negate */, typs[resultIdx].Family() == types.TupleFamily,
+		)
 		return op, resultIdx, typs, internalMemUsed, err
 	case *tree.IsNotNullExpr:
 		op, resultIdx, typs, internalMemUsed, err = planProjectionOperators(
 			ctx, evalCtx, t.TypedInnerExpr(), columnTypes, input, acc, factory,
 		)
-		op = colexec.NewIsNullSelOp(op, resultIdx, true)
+		op = colexec.NewIsNullSelOp(
+			op, resultIdx, true /* negate */, typs[resultIdx].Family() == types.TupleFamily,
+		)
 		return op, resultIdx, typs, internalMemUsed, err
 	case *tree.ComparisonExpr:
 		cmpOp := t.Operator
@@ -1465,10 +1469,11 @@ func planSelectionOperators(
 			case tree.In, tree.NotIn:
 				negate := cmpOp == tree.NotIn
 				datumTuple, ok := tree.AsDTuple(constArg)
-				if !ok {
+				if !ok || tupleContainsTuples(datumTuple) {
 					// Optimized IN operator is supported only on constant
-					// expressions, so we fallback to the default comparison
-					// operator.
+					// expressions that don't contain tuples (because tuples
+					// require special null-handling logic), so we fallback to
+					// the default comparison operator.
 					break
 				}
 				op, err = colexec.GetInOperator(lTyp, leftOp, leftIdx, datumTuple, negate)
@@ -1484,7 +1489,7 @@ func planSelectionOperators(
 				// tuples). Therefore, negate when the operator is IS DISTINCT
 				// FROM NULL.
 				negate := cmpOp == tree.IsDistinctFrom
-				op = colexec.NewIsNullSelOp(leftOp, leftIdx, negate)
+				op = colexec.NewIsNullSelOp(leftOp, leftIdx, negate, false /* isTupleNull */)
 			}
 			if op == nil {
 				// op hasn't been created yet, so let's try the constructor for
@@ -1561,6 +1566,20 @@ func planProjectionOperators(
 	acc *mon.BoundAccount,
 	factory coldata.ColumnFactory,
 ) (op colexecbase.Operator, resultIdx int, typs []*types.T, internalMemUsed int, err error) {
+	// projectDatum is a helper function that adds a new constant projection
+	// operator for the given datum. typs are updated accordingly.
+	projectDatum := func(datum tree.Datum) (colexecbase.Operator, error) {
+		resultIdx = len(columnTypes)
+		datumType := datum.ResolvedType()
+		typs = appendOneType(columnTypes, datumType)
+		if datumType.Family() == types.UnknownFamily {
+			// We handle Unknown type by planning a special constant null
+			// operator.
+			return colexec.NewConstNullOp(colmem.NewAllocator(ctx, acc, factory), input, resultIdx), nil
+		}
+		constVal := colexec.GetDatumToPhysicalFn(datumType)(datum)
+		return colexec.NewConstOp(colmem.NewAllocator(ctx, acc, factory), input, datumType, constVal, resultIdx)
+	}
 	resultIdx = -1
 	switch t := expr.(type) {
 	case *tree.IndexedVar:
@@ -1579,7 +1598,6 @@ func planProjectionOperators(
 			columnTypes, input, acc, factory, t.Fn.Fn, nil, /* cmpExpr */
 		)
 	case *tree.IsNullExpr:
-		t.TypedInnerExpr()
 		return planIsNullProjectionOp(ctx, evalCtx, t.ResolvedType(), t.TypedInnerExpr(), columnTypes, input, acc, false /* negate */, factory)
 	case *tree.IsNotNullExpr:
 		return planIsNullProjectionOp(ctx, evalCtx, t.ResolvedType(), t.TypedInnerExpr(), columnTypes, input, acc, true /* negate */, factory)
@@ -1622,21 +1640,46 @@ func planProjectionOperators(
 		typs = appendOneType(typs, t.ResolvedType())
 		return op, resultIdx, typs, internalMemUsed, err
 	case tree.Datum:
-		datumType := t.ResolvedType()
-		resultIdx = len(columnTypes)
-		typs = appendOneType(columnTypes, datumType)
-		if datumType.Family() == types.UnknownFamily {
-			// We handle Unknown type by planning a special constant null
-			// operator.
-			op = colexec.NewConstNullOp(colmem.NewAllocator(ctx, acc, factory), input, resultIdx)
-			return op, resultIdx, typs, internalMemUsed, nil
+		op, err = projectDatum(t)
+		return op, resultIdx, typs, internalMemUsed, err
+	case *tree.Tuple:
+		isConstTuple := true
+		for _, expr := range t.Exprs {
+			if _, isDatum := expr.(tree.Datum); !isDatum {
+				isConstTuple = false
+				break
+			}
 		}
-		constVal := colexec.GetDatumToPhysicalFn(datumType)(t)
-		op, err := colexec.NewConstOp(colmem.NewAllocator(ctx, acc, factory), input, datumType, constVal, resultIdx)
-		if err != nil {
-			return nil, resultIdx, typs, internalMemUsed, err
+		if isConstTuple {
+			// Tuple expression is a constant, so we can evaluate it and
+			// project the resulting datum.
+			tuple, err := t.Eval(evalCtx)
+			if err != nil {
+				return nil, resultIdx, typs, internalMemUsed, err
+			}
+			op, err = projectDatum(tuple)
+			return op, resultIdx, typs, internalMemUsed, err
 		}
-		return op, resultIdx, typs, internalMemUsed, nil
+		outputType := t.ResolvedType()
+		typs = make([]*types.T, len(columnTypes))
+		copy(typs, columnTypes)
+		tupleContentsIdxs := make([]int, len(t.Exprs))
+		for i, expr := range t.Exprs {
+			var memUsed int
+			input, tupleContentsIdxs[i], typs, memUsed, err = planProjectionOperators(
+				ctx, evalCtx, expr.(tree.TypedExpr), typs, input, acc, factory,
+			)
+			internalMemUsed += memUsed
+			if err != nil {
+				return nil, resultIdx, typs, internalMemUsed, err
+			}
+		}
+		resultIdx = len(typs)
+		op = colexec.NewTupleProjOp(
+			colmem.NewAllocator(ctx, acc, factory), typs, tupleContentsIdxs, outputType, input, resultIdx,
+		)
+		typs = appendOneType(typs, outputType)
+		return op, resultIdx, typs, internalMemUsed, err
 	case *tree.CaseExpr:
 		if t.Expr != nil {
 			return nil, resultIdx, typs, internalMemUsed, errors.New("CASE <expr> WHEN expressions unsupported")
@@ -1793,6 +1836,7 @@ func planProjectionExpr(
 	if err := checkSupportedProjectionExpr(left, right); err != nil {
 		return nil, resultIdx, typs, internalMemUsed, err
 	}
+	allocator := colmem.NewAllocator(ctx, acc, factory)
 	resultIdx = -1
 	// There are 3 cases. Either the left is constant, the right is constant,
 	// or neither are constant.
@@ -1812,8 +1856,8 @@ func planProjectionExpr(
 		// The projection result will be outputted to a new column which is appended
 		// to the input batch.
 		op, err = colexec.GetProjectionLConstOperator(
-			colmem.NewAllocator(ctx, acc, factory), typs, left.ResolvedType(), outputType,
-			projOp, input, rightIdx, lConstArg, resultIdx, evalCtx, binFn, cmpExpr,
+			allocator, typs, left.ResolvedType(), outputType, projOp, input,
+			rightIdx, lConstArg, resultIdx, evalCtx, binFn, cmpExpr,
 		)
 	} else {
 		var (
@@ -1827,6 +1871,18 @@ func planProjectionExpr(
 			return nil, resultIdx, typs, internalMemUsed, err
 		}
 		internalMemUsed += internalMemUsedLeft
+		// Note that this check exists only for testing purposes. On the
+		// regular workloads, we expect that tree.Tuples have already been
+		// pre-evaluated. We also don't need fully-fledged planning as we have
+		// for tree.Tuple in planProjectionOperators because the tests only
+		// use constant tuples.
+		if tuple, ok := right.(*tree.Tuple); ok {
+			tupleDatum, err := tuple.Eval(evalCtx)
+			if err != nil {
+				return nil, resultIdx, typs, internalMemUsed, err
+			}
+			right = tupleDatum
+		}
 		if rConstArg, rConst := right.(tree.Datum); rConst {
 			// Case 2: The right is constant.
 			// The projection result will be outputted to a new column which is appended
@@ -1836,21 +1892,21 @@ func planProjectionExpr(
 			case tree.Like, tree.NotLike:
 				negate := projOp == tree.NotLike
 				op, err = colexec.GetLikeProjectionOperator(
-					colmem.NewAllocator(ctx, acc, factory), evalCtx, input, leftIdx, resultIdx,
+					allocator, evalCtx, input, leftIdx, resultIdx,
 					string(tree.MustBeDString(rConstArg)), negate,
 				)
 			case tree.In, tree.NotIn:
 				negate := projOp == tree.NotIn
 				datumTuple, ok := tree.AsDTuple(rConstArg)
-				if !ok {
+				if !ok || tupleContainsTuples(datumTuple) {
 					// Optimized IN operator is supported only on constant
-					// expressions, so we fallback to the default comparison
-					// operator.
+					// expressions that don't contain tuples (because tuples
+					// require special null-handling logic), so we fallback to
+					// the default comparison operator.
 					break
 				}
 				op, err = colexec.GetInProjectionOperator(
-					colmem.NewAllocator(ctx, acc, factory), typs[leftIdx], input, leftIdx,
-					resultIdx, datumTuple, negate,
+					allocator, typs[leftIdx], input, leftIdx, resultIdx, datumTuple, negate,
 				)
 			case tree.IsDistinctFrom, tree.IsNotDistinctFrom:
 				if right != tree.DNull {
@@ -1862,14 +1918,16 @@ func planProjectionExpr(
 				// IS NULL is replaced with IS NOT DISTINCT FROM NULL, so we want to
 				// negate when IS DISTINCT FROM is used.
 				negate := projOp == tree.IsDistinctFrom
-				op = colexec.NewIsNullProjOp(colmem.NewAllocator(ctx, acc, factory), input, leftIdx, resultIdx, negate)
+				op = colexec.NewIsNullProjOp(
+					allocator, input, leftIdx, resultIdx, negate, false, /* isTupleNull */
+				)
 			}
 			if op == nil {
 				// op hasn't been created yet, so let's try the constructor for
 				// all other projection operators.
 				op, err = colexec.GetProjectionRConstOperator(
-					colmem.NewAllocator(ctx, acc, factory), typs, right.ResolvedType(), outputType,
-					projOp, input, leftIdx, rConstArg, resultIdx, evalCtx, binFn, cmpExpr,
+					allocator, typs, right.ResolvedType(), outputType, projOp,
+					input, leftIdx, rConstArg, resultIdx, evalCtx, binFn, cmpExpr,
 				)
 			}
 		} else {
@@ -1887,8 +1945,8 @@ func planProjectionExpr(
 			internalMemUsed += internalMemUsedRight
 			resultIdx = len(typs)
 			op, err = colexec.GetProjectionOperator(
-				colmem.NewAllocator(ctx, acc, factory), typs, outputType, projOp,
-				input, leftIdx, rightIdx, resultIdx, evalCtx, binFn, cmpExpr,
+				allocator, typs, outputType, projOp, input, leftIdx, rightIdx,
+				resultIdx, evalCtx, binFn, cmpExpr,
 			)
 		}
 	}
@@ -1986,7 +2044,10 @@ func planIsNullProjectionOp(
 		ctx, evalCtx, expr, columnTypes, input, acc, factory,
 	)
 	outputIdx := len(typs)
-	op = colexec.NewIsNullProjOp(colmem.NewAllocator(ctx, acc, factory), op, resultIdx, outputIdx, negate)
+	isTupleNull := typs[resultIdx].Family() == types.TupleFamily
+	op = colexec.NewIsNullProjOp(
+		colmem.NewAllocator(ctx, acc, factory), op, resultIdx, outputIdx, negate, isTupleNull,
+	)
 	typs = appendOneType(typs, outputType)
 	return op, outputIdx, typs, internalMemUsed, err
 }
@@ -2000,4 +2061,13 @@ func appendOneType(typs []*types.T, t *types.T) []*types.T {
 	copy(newTyps, typs)
 	newTyps[len(newTyps)-1] = t
 	return newTyps
+}
+
+func tupleContainsTuples(tuple *tree.DTuple) bool {
+	for _, typ := range tuple.ResolvedType().TupleContents() {
+		if typ.Family() == types.TupleFamily {
+			return true
+		}
+	}
+	return false
 }