--- a/pkg/sql/colexec/utils.go
+++ b/pkg/sql/colexec/utils.go
@@ -43,7 +43,7 @@ var (
 // before the inlined overloaded code.
 type overloadHelper struct {
 	tmpDec1, tmpDec2 apd.Decimal
-	binFn            *tree.BinOp
+	binFn            tree.TwoArgFn
 	evalCtx          *tree.EvalContext
 }