… pkg/sql/colexec/aggregators_test.go
… pkg/sql/colexec/builtin_funcs_test.go
… pkg/sql/colexec/default_agg_test.go
… pkg/sql/colexec/default_cmp_expr.eg.go
… pkg/sql/colexec/default_cmp_op_test.go
… pkg/sql/colexec/default_cmp_proj_ops.eg.go
… pkg/sql/colexec/default_cmp_sel_ops.eg.go
… pkg/sql/colexec/default_cmp_sel_ops_tmpl.go
… pkg/sql/colexec/distinct_test.go
… pkg/sql/colexec/execgen/cmd/execgen/default_cmp_expr_gen.go
… pkg/sql/colexec/execgen/cmd/execgen/default_cmp_sel_ops_gen.go
… pkg/sql/colexec/execgen/cmd/execgen/is_null_ops_gen.go
… pkg/sql/colexec/is_null_ops.eg.go
… pkg/sql/colexec/is_null_ops.go
… pkg/sql/colexec/is_null_ops_tmpl.go
… pkg/sql/colexec/mergejoiner_test.go
… pkg/sql/colexec/proj_const_left_ops.eg.go
… pkg/sql/colexec/proj_const_ops_tmpl.go
… pkg/sql/colexec/proj_const_right_ops.eg.go
… pkg/sql/colexec/proj_non_const_ops.eg.go
… pkg/sql/colexec/proj_non_const_ops_tmpl.go
… pkg/sql/colexec/projection_ops_test.go
… pkg/sql/colexec/select_in_test.go
… pkg/sql/colexec/selection_ops.eg.go
… pkg/sql/colexec/selection_ops_test.go
… pkg/sql/colexec/selection_ops_tmpl.go
… pkg/sql/colexec/tuple_proj_op.go
… pkg/sql/colexec/utils_test.go
… pkg/sql/logictest/testdata/logic_test/vectorize_local
… pkg/sql/sem/tree/compare_test.go