@yuzefovich 2020-08-05
This shows 4x improvement over the wrapped rowexec noop processor.
- before
```
BenchmarkDefaultComparisonOp/IS_DISTINCT_FROM/useSel=false/hasNulls=false-16         	    2587	    449706 ns/op	  36.43 MB/s
BenchmarkDefaultComparisonOp/IS_DISTINCT_FROM/useSel=false/hasNulls=true-16          	    2665	    431585 ns/op	  37.96 MB/s
BenchmarkDefaultComparisonOp/IS_DISTINCT_FROM/useSel=true/hasNulls=false-16          	    2799	    435340 ns/op	  37.63 MB/s
BenchmarkDefaultComparisonOp/IS_DISTINCT_FROM/useSel=true/hasNulls=true-16           	    2900	    429555 ns/op	  38.14 MB/s
```
- after
```
BenchmarkDefaultComparisonOp/IS_DISTINCT_FROM/useSel=false/hasNulls=false-16         	    8319	    136810 ns/op	 119.76 MB/s
BenchmarkDefaultComparisonOp/IS_DISTINCT_FROM/useSel=false/hasNulls=true-16          	    8220	    139109 ns/op	 117.78 MB/s
BenchmarkDefaultComparisonOp/IS_DISTINCT_FROM/useSel=true/hasNulls=false-16          	    8780	    141413 ns/op	 115.86 MB/s
BenchmarkDefaultComparisonOp/IS_DISTINCT_FROM/useSel=true/hasNulls=true-16           	    8679	    139692 ns/op	 117.29 MB/s
```